Follow redirects immediately
Instead of storing redirects in the urls table as a new url, we'll now
follow them directly. This is faster, and also avoids the potential to
get caught in a loop due to normalization (imagine the server removes a
trailing slash and we add one).

Up to five levels of redirects are followed.