Store hostname with port in urls table
For all intents and purposes, we should consider the port a part of the
hostname. If there are different ports on the same host, those are
different capsules. They are going to have different robots.txt files,
for example.

Some failures, like DNS errors, are going to be shared between them of
course, but overall, we'd better consider them separate hosts.