Use new url in redirects for resolving relatives
When a redirect happened while visiting a page, we were using the
original url to resolve relative urls in the page, which results in
incorrect urls. This could also result in infinitely expanding urls. For
example say gemini://example.org/foo would redirect to
gemini://example.org/ and the page we get there contains a relative link
to /bar. We would resolve this to gemini://example.org/foo/bar. If this
page redirects to the same url as before, we would be expanding that url
again and again, each time adding /bar to the end.