Improve robots.txt handling
We now properly backoff from fetching robots.txt when there's an error.
We also store robots.txt in the database for a while.

Database related code was also improved. In the crawler, we now just
create one sql.DB instance and use it across all goroutines (which is
correct, and also recommended).